requirements.txt
sanic
sanic-ext
pydantic
httpx[http2]
python-dotenv
redis
asyncpg
celery
uvicorn 